requirements.txt
nltk>=3.8.1
requests>=2.31.0
beautifulsoup4>=4.12.2
PyPDF2>=3.0.1
python-docx>=0.8.11
streamlit>=1.29.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0